% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{Transcription}
\alias{Transcription}
\alias{VerboseTranscription}
\title{Transcription object
Represents a transcription response returned by model, based on the provided input.}
\description{
Transcription object
Represents a transcription response returned by model, based on the provided input.

Transcription object
Represents a transcription response returned by model, based on the provided input.

Verbose transcription object
Represents a verbose json transcription response returned by model, based on the provided input.

Verbose transcription object
Represents a verbose json transcription response returned by model, based on the provided input.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text}}{Character. The transcribed text.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Transcription-new}{\code{Transcription$new()}}
\item \href{#method-Transcription-clone}{\code{Transcription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transcription-new"></a>}}
\if{latex}{\out{\hypertarget{method-Transcription-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Transcription object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transcription$new(..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused.}

\item{\code{resp}}{List. The response object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Transcription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Transcription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transcription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{language}}{Character. The language of the input audio.}

\item{\code{duration}}{Character. The duration of the input audio.}

\item{\code{text}}{Character. The transcribed text.}

\item{\code{words}}{List. Extracted words and their corresponding timestamps.}

\item{\code{segments}}{List. Segments of the transcribed text and their corresponding details.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VerboseTranscription-new}{\code{VerboseTranscription$new()}}
\item \href{#method-VerboseTranscription-clone}{\code{VerboseTranscription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerboseTranscription-new"></a>}}
\if{latex}{\out{\hypertarget{method-VerboseTranscription-new}{}}}
\subsection{Method \code{new()}}{
Initialize the VerboseTranscription object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerboseTranscription$new(..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused.}

\item{\code{resp}}{List. The response object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VerboseTranscription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VerboseTranscription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VerboseTranscription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
