% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{Conversation}
\alias{Conversation}
\title{Conversation R6 Class}
\description{
An R6 class representing a Conversation in the OpenAI API.
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{Conversation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The conversation ID.}

\item{\code{object}}{The object type, which is "conversation".}

\item{\code{created_at}}{The creation time of the conversation, in epoch seconds.}

\item{\code{metadata}}{Metadata associated with the conversation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Conversation-new}{\code{Conversation$new()}}
\item \href{#method-Conversation-retrieve}{\code{Conversation$retrieve()}}
\item \href{#method-Conversation-update}{\code{Conversation$update()}}
\item \href{#method-Conversation-delete}{\code{Conversation$delete()}}
\item \href{#method-Conversation-list_items}{\code{Conversation$list_items()}}
\item \href{#method-Conversation-create_items}{\code{Conversation$create_items()}}
\item \href{#method-Conversation-retrieve_item}{\code{Conversation$retrieve_item()}}
\item \href{#method-Conversation-delete_item}{\code{Conversation$delete_item()}}
\item \href{#method-Conversation-clone}{\code{Conversation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Conversation object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$new(conversation_id = NULL, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conversation_id}}{Character. The ID of the conversation.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}

\item{\code{resp}}{A list. The object returned by the OpenAI API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve a fresh copy of the conversation from the API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-update"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-update}{}}}
\subsection{Method \code{update()}}{
Update the conversation's metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$update(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the conversation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-list_items"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-list_items}{}}}
\subsection{Method \code{list_items()}}{
List items in the conversation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$list_items(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-create_items"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-create_items}{}}}
\subsection{Method \code{create_items()}}{
Add items to the conversation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$create_items(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-retrieve_item"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-retrieve_item}{}}}
\subsection{Method \code{retrieve_item()}}{
Retrieve a specific item from the conversation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$retrieve_item(item_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item_id}}{Character. The ID of a specific conversation item.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-delete_item"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-delete_item}{}}}
\subsection{Method \code{delete_item()}}{
Delete a specific item from the conversation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$delete_item(item_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item_id}}{Character. The ID of a specific conversation item.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Conversation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Conversation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Conversation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
