% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{Embedding}
\alias{Embedding}
\title{R6 class representing an Embedding in the OpenAI API}
\description{
R6 class representing an Embedding in the OpenAI API

R6 class representing an Embedding in the OpenAI API
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{embedding}}{The embedding vector.}

\item{\code{index}}{The index of the embedding in the list of embeddings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Embedding-new}{\code{Embedding$new()}}
\item \href{#method-Embedding-print}{\code{Embedding$print()}}
\item \href{#method-Embedding-clone}{\code{Embedding$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Embedding-new"></a>}}
\if{latex}{\out{\hypertarget{method-Embedding-new}{}}}
\subsection{Method \code{new()}}{
Initialize an Embedding object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Embedding$new(input, model, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Character. Input text to embed.}

\item{\code{model}}{Character. ID of the model used to generate the embedding.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}

\item{\code{resp}}{A list. The response object from the OpenAI API containing details of the embedding.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of an Embedding object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Embedding-print"></a>}}
\if{latex}{\out{\hypertarget{method-Embedding-print}{}}}
\subsection{Method \code{print()}}{
Print the Embedding object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Embedding$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Embedding-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Embedding-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Embedding$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
