% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{File}
\alias{File}
\title{OpenAI File Interface}
\description{
OpenAI File Interface

OpenAI File Interface
}
\details{
This class provides methods to interact with files in the OpenAI API.
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{File}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The ID of the file.}

\item{\code{bytes}}{The size of the file in bytes.}

\item{\code{created_at}}{The creation time of the file.}

\item{\code{filename}}{The name of the file.}

\item{\code{purpose}}{The purpose of the file.
Retrieve the file's properties from the API}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-File-new}{\code{File$new()}}
\item \href{#method-File-retrieve}{\code{File$retrieve()}}
\item \href{#method-File-delete}{\code{File$delete()}}
\item \href{#method-File-print}{\code{File$print()}}
\item \href{#method-File-clone}{\code{File$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="openaiapi" data-topic="Utils" data-id="store_response"><a href='../../openaiapi/html/Utils.html#method-Utils-store_response'><code>openaiapi::Utils$store_response()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-new"></a>}}
\if{latex}{\out{\hypertarget{method-File-new}{}}}
\subsection{Method \code{new()}}{
Initializes the File object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$new(file_id = NULL, path = NULL, ..., resp = NULL, .async = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{The ID of the file to initialize.}

\item{\code{path}}{The path to the file to upload.}

\item{\code{...}}{Additional arguments passed to the API functions.}

\item{\code{resp}}{A list containing the file properties from the API response.}

\item{\code{.async}}{Logical. If TRUE, the function will return a promise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the File class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-File-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$retrieve()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The up-to-date File instance.
Delete the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-delete"></a>}}
\if{latex}{\out{\hypertarget{method-File-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$delete()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Print the file's details
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-print"></a>}}
\if{latex}{\out{\hypertarget{method-File-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (unused).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The File instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-clone"></a>}}
\if{latex}{\out{\hypertarget{method-File-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
