% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_api}
\alias{files_api}
\alias{oai_upload_file}
\alias{oai_list_files}
\alias{oai_retrieve_file}
\alias{oai_delete_file}
\alias{oai_retrieve_file_content}
\title{Files API}
\usage{
oai_upload_file(
  path,
  purpose = c("assistants", "vision", "batch", "fine-tune"),
  name = NULL,
  .classify_response = TRUE,
  .async = FALSE
)

oai_list_files(
  purpose = NULL,
  limit = NULL,
  order = NULL,
  after = NULL,
  .classify_response = TRUE
)

oai_retrieve_file(file_id, .classify_response = TRUE, .async = FALSE)

oai_delete_file(file_id, .async = FALSE)

oai_retrieve_file_content(file_id, path, .async = FALSE)
}
\arguments{
\item{path}{Character. The path to the file to upload.}

\item{purpose}{Character. Optional. Only list files with the specified purpose.}

\item{name}{Character. Optional. The name of the file. If not provided, the file name will be used.}

\item{.classify_response}{Logical. Whether to classify the response as an R6 object.}

\item{.async}{Logical. If \code{TRUE}, the request is performed asynchronously.}

\item{limit}{Integer. Optional. The maximum number of files to list. Defaults to 10,000.}

\item{order}{Character. Optional. The order to list files. Can be "asc" or "desc".}

\item{after}{Character. Optional. The ID of the file to start the list from.}

\item{file_id}{Character. The ID of the file to delete.}
}
\value{
A File R6 object

File R6 object

File R6 object

Deletion status

The response from the API.
}
\description{
Files are used to upload documents that can be used with features like Assistants, Fine-tuning, and Batch API.

\itemize{
\item \code{oai_upload_file()} Upload a file to OpenAI.
}

\itemize{
\item \code{oai_list_files()} List all files uploaded to OpenAI.
}

\itemize{
\item \code{oai_retrieve_file()} Retrieve a specific file from OpenAI.
}

\itemize{
\item \code{oai_retrieve_file_content()} Retrieve the content of a specific file from OpenAI.
}
}
