% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{fine_tuning_api}
\alias{fine_tuning_api}
\alias{oai_create_fine_tuning_job}
\alias{oai_list_fine_tuning_jobs}
\alias{oai_list_fine_tuning_events}
\alias{oai_list_fine_tuning_checkpoints}
\alias{oai_get_fine_tuning_job}
\alias{oai_cancel_fine_tuning}
\title{Fine-tuning API}
\usage{
oai_create_fine_tuning_job(
  model,
  training_file,
  hyperparameters = NULL,
  suffix = NULL,
  validation_file = NULL,
  integrations = NULL,
  seed = NULL,
  .classify_response = TRUE,
  .async = FALSE
)

oai_list_fine_tuning_jobs(
  after = NULL,
  limit = NULL,
  .classify_response = TRUE
)

oai_list_fine_tuning_events(
  fine_tuning_job_id,
  after = NULL,
  limit = NULL,
  .classify_response = TRUE
)

oai_list_fine_tuning_checkpoints(
  fine_tuning_job_id,
  after = NULL,
  limit = NULL,
  .classify_response = TRUE
)

oai_get_fine_tuning_job(
  fine_tuning_job_id,
  .classify_response = TRUE,
  .async = FALSE
)

oai_cancel_fine_tuning(
  fine_tuning_job_id,
  .classify_response = TRUE,
  .async = FALSE
)
}
\arguments{
\item{model}{The name of the model to fine-tune. You can select one of the supported models.}

\item{training_file}{The ID of an uploaded file that contains training data.}

\item{hyperparameters}{The hyperparameters used for the fine-tuning job.}

\item{suffix}{A string of up to 64 characters that will be added to your fine-tuned model name.}

\item{validation_file}{The ID of an uploaded file that contains validation data.}

\item{integrations}{A list of integrations to enable for your fine-tuning job.}

\item{seed}{The seed controls the reproducibility of the job. Passing in the same seed and job parameters should produce the same results, but may differ in rare cases. If a seed is not specified, one will be generated for you.}

\item{.classify_response}{Logical. If \code{TRUE} (default), the response is classified as an R6 object. If \code{FALSE}, the response is returned as a list.}

\item{.async}{Logical. If \code{TRUE}, the request is performed asynchronously.}

\item{after}{Identifier for the last checkpoint ID from the previous pagination request.}

\item{limit}{Number of checkpoints to retrieve.}

\item{fine_tuning_job_id}{The ID of the fine-tuning job.}
}
\value{
A FineTuningJob R6 object.

\code{oai_list_fine_tuning_jobs()} returns a list of FineTuningJob R6 objects.

\code{oai_list_fine_tuning_events()} returns a list of FineTuningEvent R6 objects.

\code{oai_list_fine_tuning_checkpoints()} returns a list of FineTuningCheckpoint R6 objects.
}
\description{
Manage fine-tuning jobs to tailor a model to your specific training data.

\itemize{
\item \code{oai_create_fine_tuning_job()} creates a fine-tuning job.
}

\itemize{
\item \code{oai_list_fine_tuning_jobs()} lists fine-tuning jobs.
}

\itemize{
\item \code{oai_list_fine_tuning_events()} lists fine-tuning events.
}

\itemize{
\item \code{oai_list_fine_tuning_checkpoints()} lists fine-tuning checkpoints.
}

\itemize{
\item \code{oai_get_fine_tuning_job()} retrieves a fine-tuning job.
}

\itemize{
\item \code{oa_cancel_fine_tuning()} cancels a fine-tuning job.
}
}
\seealso{
\code{FineTuningJob}
}
