% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{FineTuningJob}
\alias{FineTuningJob}
\title{FineTuningJob R6 class}
\description{
FineTuningJob R6 class

FineTuningJob R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The ID of the fine-tuning job.}

\item{\code{created_at}}{The time the fine-tuning job was created.}

\item{\code{error}}{The error message, if any.}

\item{\code{fine_tuned_model}}{The name of the fine-tuned model.}

\item{\code{finished_at}}{The time the fine-tuning job finished.}

\item{\code{hyperparameters}}{The hyperparameters used for the fine-tuning job.}

\item{\code{model}}{The name of the model to fine-tune.}

\item{\code{object}}{The object type.}

\item{\code{organization_id}}{The ID of the organization.}

\item{\code{result_files}}{The result files.}

\item{\code{status}}{The status of the fine-tuning job.}

\item{\code{trained_tokens}}{The number of tokens trained.}

\item{\code{training_file}}{The ID of the training file.}

\item{\code{validation_file}}{The ID of the validation file.}

\item{\code{integrations}}{The integrations enabled for the fine-tuning job.}

\item{\code{seed}}{The seed used for the fine-tuning job.}

\item{\code{estimated_finish}}{The estimated finish time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FineTuningJob-new}{\code{FineTuningJob$new()}}
\item \href{#method-FineTuningJob-list_events}{\code{FineTuningJob$list_events()}}
\item \href{#method-FineTuningJob-list_checkpoints}{\code{FineTuningJob$list_checkpoints()}}
\item \href{#method-FineTuningJob-retrieve}{\code{FineTuningJob$retrieve()}}
\item \href{#method-FineTuningJob-cancel}{\code{FineTuningJob$cancel()}}
\item \href{#method-FineTuningJob-clone}{\code{FineTuningJob$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-new}{}}}
\subsection{Method \code{new()}}{
Initialize a FineTuningJob object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$new(model = NULL, training_file = NULL, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The name of the model to fine-tune.}

\item{\code{training_file}}{The ID of an uploaded file that contains training data.}

\item{\code{...}}{Additional parameters passed to the API functions.}

\item{\code{resp}}{The response from the API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-list_events"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-list_events}{}}}
\subsection{Method \code{list_events()}}{
List fine-tuning events.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$list_events(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-list_checkpoints"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-list_checkpoints}{}}}
\subsection{Method \code{list_checkpoints()}}{
List fine-tuning checkpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$list_checkpoints(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the fine-tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancel the fine-tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$cancel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineTuningJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FineTuningJob-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineTuningJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
