% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{Message}
\alias{Message}
\title{Message R6 class}
\description{
Message R6 class

Message R6 class
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{Message}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Character. The ID of the message.}

\item{\code{created_at}}{POSIXct. The time the message was created.}

\item{\code{thread_id}}{Character. The ID of the thread this message belongs to.}

\item{\code{status}}{Character. The status of the message.}

\item{\code{incomplete_details}}{List. Details about the message if it is incomplete.}

\item{\code{completed_at}}{POSIXct. The time the message was completed.}

\item{\code{incomplete_at}}{POSIXct. The time the message was marked as incomplete.}

\item{\code{role}}{Character. The role of the sender.}

\item{\code{content}}{List. The content of the message.}

\item{\code{assistant_id}}{Character. If applicable, the ID of the assistant that authored this message.}

\item{\code{run_id}}{Character. The ID of the run associated with the creation of this message. Value is null when messages are created manually using the create message or create thread endpoints.}

\item{\code{attachments}}{List. Attachments to the message.}

\item{\code{metadata}}{List. Metadata associated with the message.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Message-new}{\code{Message$new()}}
\item \href{#method-Message-retrieve}{\code{Message$retrieve()}}
\item \href{#method-Message-modify}{\code{Message$modify()}}
\item \href{#method-Message-delete}{\code{Message$delete()}}
\item \href{#method-Message-content_text}{\code{Message$content_text()}}
\item \href{#method-Message-thread}{\code{Message$thread()}}
\item \href{#method-Message-assistant}{\code{Message$assistant()}}
\item \href{#method-Message-clone}{\code{Message$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-new"></a>}}
\if{latex}{\out{\hypertarget{method-Message-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Message object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$new(
  message_id = NULL,
  thread_id = NULL,
  ...,
  resp = NULL,
  .async = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message_id}}{Character. The ID of the message.}

\item{\code{thread_id}}{Character. The ID of the thread.}

\item{\code{...}}{Additional parameters passed to the API call.}

\item{\code{resp}}{List. The response from the OpenAI API.}

\item{\code{.async}}{Logical. If TRUE, the API call will be asynchronous.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-Message-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-modify"></a>}}
\if{latex}{\out{\hypertarget{method-Message-modify}{}}}
\subsection{Method \code{modify()}}{
Modify the message metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$modify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to the API call.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Message-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-content_text"></a>}}
\if{latex}{\out{\hypertarget{method-Message-content_text}{}}}
\subsection{Method \code{content_text()}}{
Get the text content of the message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$content_text()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-thread"></a>}}
\if{latex}{\out{\hypertarget{method-Message-thread}{}}}
\subsection{Method \code{thread()}}{
Get the thread this message belongs to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$thread()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-assistant"></a>}}
\if{latex}{\out{\hypertarget{method-Message-assistant}{}}}
\subsection{Method \code{assistant()}}{
Get the assistant if applicable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$assistant()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Message-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Message-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Message$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
