% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R, R/helpers.R
\name{oai_create_message}
\alias{oai_create_message}
\alias{oai_attachment}
\alias{message_api}
\alias{oai_list_messages}
\alias{oai_retrieve_message}
\alias{oai_modify_message}
\alias{oai_delete_message}
\title{A helper function for the \code{attachment} argument in \code{oai_create_message()} and \code{Message} objects.}
\usage{
oai_create_message(
  thread_id,
  content,
  role = c("user", "assistant"),
  attachments = NULL,
  metadata = NULL,
  .classify_response = TRUE,
  .async = FALSE
)

oai_attachment(file_id, tools = c("code_interpreter", "file_search"))

oai_list_messages(
  thread_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  .classify_response = TRUE,
  .async = FALSE
)

oai_retrieve_message(
  thread_id,
  message_id,
  .classify_response = TRUE,
  .async = FALSE
)

oai_modify_message(
  thread_id,
  message_id,
  metadata = NULL,
  .classify_response = TRUE,
  .async = FALSE
)

oai_delete_message(thread_id, message_id, .async = FALSE)
}
\arguments{
\item{thread_id}{Character. The ID of the thread.}

\item{content}{Character. The content of the message.}

\item{role}{Character. The role of the sender.}

\item{attachments}{List or NULL. Optional. Attachments to include with the message.}

\item{metadata}{List or NULL. Optional. Metadata to include with the message.}

\item{.classify_response}{Logical. If \code{TRUE} (default), the response is classified as an R6 object. If \code{FALSE}, the response is returned as a list.}

\item{.async}{Logical. If \code{TRUE}, the request is performed asynchronously.}

\item{file_id}{Character. The ID of the file.}

\item{tools}{Character. The tools to use with the attachment.}

\item{limit}{Integer. Optional. A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 20.}

\item{order}{Character. Optional. Sort order by the created_at timestamp of the objects. \code{"asc"} for ascending order and \code{"desc"} for descending order.}

\item{after}{Character. Optional. A cursor for use in pagination. \code{after} is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include \code{after = "obj_foo"} in order to fetch the next page of the list.}

\item{before}{Character. Optional. A cursor for use in pagination. before is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include \code{before = "obj_foo"} in order to fetch the previous page of the list.}

\item{message_id}{Character. The ID of the message.}
}
\value{
A Message R6 object.

\code{oai_list_messages()} returns a list of messages in a thread.

\code{oai_delete_message()} returns the deletion status.
}
\description{
\itemize{
\item \code{oai_create_message()} Create a message in a thread.
}

Manage messages within threads

\itemize{
\item \code{oai_list_messages()} List messages in a thread.
}

\itemize{
\item \code{oai_retrieve_message()} Retrieve a message.
}

\itemize{
\item \code{oai_modify_message()} Modify a message.
}

\itemize{
\item \code{oai_delete_message()} Delete a message.
}
}
