% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{models_api}
\alias{models_api}
\alias{oai_list_models}
\alias{oai_retrieve_model}
\alias{oai_delete_model}
\title{Models API}
\usage{
oai_list_models()

oai_retrieve_model(model_id)

oai_delete_model(model_id)
}
\arguments{
\item{model_id}{Character. The ID of the model to retrieve/delete.}
}
\value{
of models
}
\description{
List and describe the various models available in the API.

\itemize{
\item \code{oai_list_models()}: Retrieve a data frame of models.
}

\itemize{
\item \code{oai_retrieve_model()}: Retrieve a model.
}

\itemize{
\item \code{oai_delete_model()}: Delete a model.
}
}
