% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{oai_function_tool}
\alias{oai_function_tool}
\alias{oai_function_tool_parameter}
\title{Define a function tool}
\usage{
oai_function_tool(
  name,
  description = NULL,
  ...,
  parameters = NULL,
  strict = TRUE
)

oai_function_tool_parameter(
  name,
  type,
  description,
  required = TRUE,
  enum = NULL,
  ...
)
}
\arguments{
\item{name}{The name of the parameter.}

\item{description}{A description of the parameter.}

\item{...}{Additional properties to be added to the parameter.}

\item{parameters}{Parameters can be given alternatively a list of \code{oai_function_tool_parameter} objects.}

\item{strict}{Whether to enable strict schema adherence when generating the function call. If set to true, the model will follow the exact schema defined in the parameters field.}

\item{type}{The type of the parameter. Possible values are "string", "number", "boolean", "array", "object", and "null".}

\item{required}{Whether the parameter is required or optional.}

\item{enum}{A list of possible values for the parameter.}
}
\description{
Define a function tool

Create a function tool parameter for \code{oai_function_tool}
}
