% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaiapi-package.R
\docType{package}
\name{openaiapi-package}
\alias{openaiapi}
\alias{openaiapi-package}
\title{openaiapi: What the Package Does}
\description{
The goal of openaiapi is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/openaiapi}
  \item \url{https://jonthegeek.github.io/openaiapi/}
  \item Report bugs at \url{https://github.com/jonthegeek/openaiapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
