% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{Run}
\alias{Run}
\title{Run R6 class}
\description{
Run R6 class

Run R6 class
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{Run}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Run ID.}

\item{\code{created_at}}{Time the run was created.}

\item{\code{thread_id}}{Thread ID.}

\item{\code{assistant_id}}{Assistant ID.}

\item{\code{status}}{Run status.}

\item{\code{required_action}}{Required action.}

\item{\code{last_error}}{Last error.}

\item{\code{started_at}}{Time the run was started.}

\item{\code{expires_at}}{Time the run expires.}

\item{\code{cancelled_at}}{Time the run was cancelled.}

\item{\code{failed_at}}{Time the run failed.}

\item{\code{completed_at}}{Time the run was completed.}

\item{\code{incomplete_details}}{Incomplete details.}

\item{\code{model}}{Model.}

\item{\code{instructions}}{Instructions.}

\item{\code{tools}}{Tools.}

\item{\code{metadata}}{Metadata.}

\item{\code{usage}}{Usage.}

\item{\code{temperature}}{Temperature.}

\item{\code{top_p}}{Top p.}

\item{\code{max_prompt_tokens}}{Maximum prompt tokens.}

\item{\code{max_completion_tokens}}{Maximum completion tokens.}

\item{\code{truncation_strategy}}{Truncation strategy.}

\item{\code{tool_choice}}{Tool choice.}

\item{\code{response_format}}{Response format.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Run-new}{\code{Run$new()}}
\item \href{#method-Run-list_steps}{\code{Run$list_steps()}}
\item \href{#method-Run-retrieve}{\code{Run$retrieve()}}
\item \href{#method-Run-retrieve_status}{\code{Run$retrieve_status()}}
\item \href{#method-Run-modify}{\code{Run$modify()}}
\item \href{#method-Run-cancel}{\code{Run$cancel()}}
\item \href{#method-Run-submit_tool_outputs}{\code{Run$submit_tool_outputs()}}
\item \href{#method-Run-wait}{\code{Run$wait()}}
\item \href{#method-Run-await}{\code{Run$await()}}
\item \href{#method-Run-do_tool_calls}{\code{Run$do_tool_calls()}}
\item \href{#method-Run-thread}{\code{Run$thread()}}
\item \href{#method-Run-assistant}{\code{Run$assistant()}}
\item \href{#method-Run-clone}{\code{Run$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-new"></a>}}
\if{latex}{\out{\hypertarget{method-Run-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Run object. If \code{thread_id} and \code{run_id} are provided, the Run object is initialized with the corresponding run. If \code{thread_id} and \code{assistant_id} are provided, a new run is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$new(
  thread_id = NULL,
  run_id = NULL,
  assistant_id = NULL,
  ...,
  resp = NULL,
  .async = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{Thread ID or Thread object}

\item{\code{run_id}}{Run ID.}

\item{\code{assistant_id}}{Assistant ID or Assistant object.}

\item{\code{...}}{Additional arguments passed to the API function.}

\item{\code{resp}}{Response object.}

\item{\code{.async}}{Logical. If TRUE, the API call will be asynchronous.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-list_steps"></a>}}
\if{latex}{\out{\hypertarget{method-Run-list_steps}{}}}
\subsection{Method \code{list_steps()}}{
Retrieve the run. The \code{...} argument is passed to \verb{oai_list_run steps()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$list_steps(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list_steps()} returns a list of \code{RunStep} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-Run-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-retrieve_status"></a>}}
\if{latex}{\out{\hypertarget{method-Run-retrieve_status}{}}}
\subsection{Method \code{retrieve_status()}}{
Retrieve the up-to-date status of the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$retrieve_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-modify"></a>}}
\if{latex}{\out{\hypertarget{method-Run-modify}{}}}
\subsection{Method \code{modify()}}{
Modify the metadata of the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$modify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-Run-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancel the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$cancel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-submit_tool_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-Run-submit_tool_outputs}{}}}
\subsection{Method \code{submit_tool_outputs()}}{
Submit tool outputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$submit_tool_outputs(tool_outputs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tool_outputs}}{Tool outputs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-wait"></a>}}
\if{latex}{\out{\hypertarget{method-Run-wait}{}}}
\subsection{Method \code{wait()}}{
Wait for the run to complete.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$wait(env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{Environment to evaluate tool calls.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-await"></a>}}
\if{latex}{\out{\hypertarget{method-Run-await}{}}}
\subsection{Method \code{await()}}{
Perform tool calls asynchronously. Returns a promise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$await(env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{Environment to evaluate tool calls.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-do_tool_calls"></a>}}
\if{latex}{\out{\hypertarget{method-Run-do_tool_calls}{}}}
\subsection{Method \code{do_tool_calls()}}{
Perform tool calls. Returns a list of tool outputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$do_tool_calls(env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{env}}{Environment to evaluate tool calls.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-thread"></a>}}
\if{latex}{\out{\hypertarget{method-Run-thread}{}}}
\subsection{Method \code{thread()}}{
Retrieve the thread of the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$thread()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-assistant"></a>}}
\if{latex}{\out{\hypertarget{method-Run-assistant}{}}}
\subsection{Method \code{assistant()}}{
Retrieve the assistant of the run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$assistant()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Run-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Run-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Run$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
