% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{Thread}
\alias{Thread}
\title{Thread R6 class}
\description{
Thread R6 class

Thread R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The ID of the thread.}

\item{\code{created_at}}{The time the thread was created.}

\item{\code{metadata}}{Additional metadata about the thread.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Thread-new}{\code{Thread$new()}}
\item \href{#method-Thread-modify}{\code{Thread$modify()}}
\item \href{#method-Thread-retrieve}{\code{Thread$retrieve()}}
\item \href{#method-Thread-delete}{\code{Thread$delete()}}
\item \href{#method-Thread-create_message}{\code{Thread$create_message()}}
\item \href{#method-Thread-list_messages}{\code{Thread$list_messages()}}
\item \href{#method-Thread-run}{\code{Thread$run()}}
\item \href{#method-Thread-list_runs}{\code{Thread$list_runs()}}
\item \href{#method-Thread-clone}{\code{Thread$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-new"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-new}{}}}
\subsection{Method \code{new()}}{
Initialize the thread. If \code{thread_id} is provided, the thread is retrieved from the API. The \code{...} argument is passed to \code{oai_create_thread()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$new(thread_id = NULL, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{The ID of the thread.}

\item{\code{...}}{Additional arguments passed to the API functions.}

\item{\code{resp}}{A response from the OpenAI API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-modify"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-modify}{}}}
\subsection{Method \code{modify()}}{
Modify the thread. The \code{...} argument is passed to \code{oai_modify_thread()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$modify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the thread. The \code{...} argument is passed to \code{oai_retrieve_thread()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$retrieve(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-create_message"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-create_message}{}}}
\subsection{Method \code{create_message()}}{
Create a new message in the thread. The \code{...} argument is passed to \code{oai_create_message()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$create_message(content, role = "user", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{The content of the message.}

\item{\code{role}}{The role of the message. Can be "user" or "assistant".}

\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{create_message()} returns the thread object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-list_messages"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-list_messages}{}}}
\subsection{Method \code{list_messages()}}{
List messages in the thread. The \code{...} argument is passed to \code{oai_list_messages()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$list_messages(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-run"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-run}{}}}
\subsection{Method \code{run()}}{
Run the thread on an assistant. The \code{...} argument is passed to \code{oai_create_run()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$run(assistant_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{The assistant to run the thread on. Can be an Assistant object or an assistant ID.}

\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-list_runs"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-list_runs}{}}}
\subsection{Method \code{list_runs()}}{
List runs of the thread. The \code{...} argument is passed to \code{oai_list_runs()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$list_runs(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Thread-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Thread-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Thread$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
