% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{thread_api}
\alias{thread_api}
\alias{oai_create_thread}
\alias{oai_modify_thread}
\alias{oai_retrieve_thread}
\alias{oai_delete_thread}
\title{Threads API}
\usage{
oai_create_thread(
  messages = NULL,
  tool_resources = NULL,
  metadata = NULL,
  .classify_response = TRUE
)

oai_modify_thread(
  thread_id,
  tool_resources = NULL,
  metadata = NULL,
  .classify_response = TRUE
)

oai_retrieve_thread(thread_id, .classify_response = TRUE, .async = FALSE)

oai_delete_thread(thread_id)
}
\arguments{
\item{messages}{A list of oai_messages to start the thread with.}

\item{tool_resources}{A list of tool resources that are available to the assistant.}

\item{metadata}{List. Optional. Set of 16 key-value pairs that can be attached to an object. This can be useful for storing additional information about the object in a structured format. Keys can be a maximum of 64 characters long and values can be a maximum of 512 characters long.}

\item{.classify_response}{Logical. If \code{TRUE} (default), the response is classified as an R6 object. If \code{FALSE}, the response is returned as a list.}

\item{thread_id}{The ID of the thread.}

\item{.async}{Logical. If \code{TRUE}, the request is performed asynchronously.}
}
\description{
Manage threads that assistants can interact with.

\itemize{
\item \code{oai_create_thread()}: Create a new thread that assistants can interact with.
}

\itemize{
\item \code{oai_modify_thread()}: Modify a thread.
}

\itemize{
\item \code{oai_retrieve_thread()}: Retrieve a thread.
}

\itemize{
\item \code{oai_delete_thread()}: Delete a thread.
}
}
