% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{oai_message}
\alias{oai_message}
\alias{oai_thread}
\alias{oai_tool_resource}
\title{A helper function for message arguments}
\usage{
oai_message(content, role = "user", attachments = NULL, metadata = NULL, ...)

oai_thread(messages, tool_resources = NULL, metadata = NULL)

oai_tool_resource(code_interpreter_file_ids = NULL, vector_store_ids = NULL)
}
\arguments{
\item{content}{Character. The content of the message.}

\item{role}{Character. The role of the sender (default is "user").}

\item{attachments}{List or NULL. Optional. Attachments to include with the message.}

\item{metadata}{List or NULL. Optional. Metadata to include with the thread.}

\item{...}{Additional values to pass to the API.}

\item{messages}{A list of oai_message objects or a character vector.}

\item{tool_resources}{List. Optional. A list of tool resources to include with the thread.}

\item{code_interpreter_file_ids, }{Character A vector of file IDs fore the code interpreter to use.}

\item{vector_store_ids, }{Character A vector of vector store IDs for the file search to use.}
}
\value{
A structured list representing the message.
}
\description{
A helper function for message arguments

A helper function for thread arguments

A helper function for tool resource arguments
}
