% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_store_file_batches.R
\name{vector_store_file_batch_api}
\alias{vector_store_file_batch_api}
\alias{oai_create_vector_store_file_batch}
\alias{oai_retrieve_vector_store_file_batch}
\alias{oai_cancel_vector_store_file_batch}
\alias{oai_list_vector_store_files_in_a_batch}
\title{Vector store file batch API}
\usage{
oai_create_vector_store_file_batch(
  vector_store_id,
  file_ids,
  chunking_strategy = NULL,
  .classify_response = TRUE
)

oai_retrieve_vector_store_file_batch(
  vector_store_id,
  batch_id,
  .classify_response = TRUE
)

oai_cancel_vector_store_file_batch(vector_store_id, batch_id)

oai_list_vector_store_files_in_a_batch(
  vector_store_id,
  batch_id,
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL,
  filter = NULL
)
}
\arguments{
\item{vector_store_id}{Character. The ID of the vector store where the file batch will be created.}

\item{file_ids}{Character. A vector of file IDs to include in the batch.}

\item{chunking_strategy}{List. Optional. Strategy for chunking data.}

\item{.classify_response}{Logical. If \code{TRUE} (default), the response is classified as an R6 object. If \code{FALSE}, the response is returned as a list.}

\item{batch_id}{Character. The ID of the batch to retrieve.}

\item{limit}{Integer. Optional. A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 20.}

\item{order}{Character. Optional. Sort order by the created_at timestamp of the objects. \code{"asc"} for ascending order and \code{"desc"} for descending order.}

\item{after}{Character. Optional. A cursor for use in pagination. \code{after} is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include \code{after = "obj_foo"} in order to fetch the next page of the list.}

\item{before}{Character. Optional. A cursor for use in pagination. before is an object ID that defines your place in the list. For instance, if you make a list request and receive 100 objects, ending with obj_foo, your subsequent call can include \code{before = "obj_foo"} in order to fetch the previous page of the list.}

\item{filter}{Character. Optional. Filter the files based on certain criteria.}
}
\value{
A VectorStoreFilesBatch R6 object.

\code{oai_list_vector_store_files_in_a_batch()} returns a list of VectorStoreFile objects.
}
\description{
Vector store file batches represent operations to add multiple files to a vector store.

\itemize{
\item \code{oai_create_vector_store_file_batch()}: Create a file batch in a vector store.
}

\itemize{
\item \code{oai_retrieve_vector_store_file_batch()}: Retrieve a specific file batch in a vector store.
}

\itemize{
\item \code{oai_delete_vector_store_file_batch()}: Delete a specific file batch in a vector store.
}

\itemize{
\item \code{oai_list_vector_store_files_in_a_batch()}: List all files in a file batch.
}
}
