% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{VectorStore}
\alias{VectorStore}
\title{R6 class for managing a Vector Store in the OpenAI API}
\description{
R6 class for managing a Vector Store in the OpenAI API

R6 class for managing a Vector Store in the OpenAI API
}
\details{
The \code{VectorStore} class provides methods to manage vector stores.
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{VectorStore}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Character. The unique identifier of the vector store.}

\item{\code{created_at}}{POSIXct. The timestamp when the vector store was created.}

\item{\code{name}}{Character. The name of the vector store.}

\item{\code{usage_bytes}}{Numeric. The total number of bytes used by the vector store.}

\item{\code{file_counts}}{Integer. The number of files in the vector store.}

\item{\code{status}}{Character. The current status of the vector store.}

\item{\code{expires_after}}{List. The duration after which the vector store will expire.}

\item{\code{expires_at}}{POSIXct. The exact timestamp when the vector store will expire.}

\item{\code{last_active_at}}{POSIXct. The last time the vector store was active.}

\item{\code{metadata}}{List. Additional metadata associated with the vector store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStore-new}{\code{VectorStore$new()}}
\item \href{#method-VectorStore-retrieve}{\code{VectorStore$retrieve()}}
\item \href{#method-VectorStore-modify}{\code{VectorStore$modify()}}
\item \href{#method-VectorStore-delete}{\code{VectorStore$delete()}}
\item \href{#method-VectorStore-create_file}{\code{VectorStore$create_file()}}
\item \href{#method-VectorStore-upload_file}{\code{VectorStore$upload_file()}}
\item \href{#method-VectorStore-list_files}{\code{VectorStore$list_files()}}
\item \href{#method-VectorStore-retrieve_file}{\code{VectorStore$retrieve_file()}}
\item \href{#method-VectorStore-delete_file}{\code{VectorStore$delete_file()}}
\item \href{#method-VectorStore-search}{\code{VectorStore$search()}}
\item \href{#method-VectorStore-create_file_batch}{\code{VectorStore$create_file_batch()}}
\item \href{#method-VectorStore-retrieve_file_batch}{\code{VectorStore$retrieve_file_batch()}}
\item \href{#method-VectorStore-cancel_file_batch}{\code{VectorStore$cancel_file_batch()}}
\item \href{#method-VectorStore-list_files_in_a_batch}{\code{VectorStore$list_files_in_a_batch()}}
\item \href{#method-VectorStore-print}{\code{VectorStore$print()}}
\item \href{#method-VectorStore-clone}{\code{VectorStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-new}{}}}
\subsection{Method \code{new()}}{
Initialize a VectorStore object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$new(vector_store_id = NULL, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector_store_id}}{Character. Optional. The ID of the vector store to initialize.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}

\item{\code{resp}}{A list. The response object from the OpenAI API containing details of the vector store.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of a VectorStore object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the vector store
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$retrieve()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The updated VectorStore object after retrieving its details from the OpenAI API.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-modify"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-modify}{}}}
\subsection{Method \code{modify()}}{
Modify the vector store. The \code{...} argument is passed to \code{oai_modify_vector_store()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$modify(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-delete"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the vector store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-create_file"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-create_file}{}}}
\subsection{Method \code{create_file()}}{
Create a file in the vector store. The \code{...} argument is passed to \code{oai_create_vector_store_file()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$create_file(file_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{Character. The ID of the file to be added to the vector store.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-upload_file"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-upload_file}{}}}
\subsection{Method \code{upload_file()}}{
Upload a file to the vector store. The \code{...} argument is passed to \code{oai_upload_file()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$upload_file(path, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character. The path to the file to upload.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-list_files}{}}}
\subsection{Method \code{list_files()}}{
List files in the vector store. The \code{...} argument is passed to \code{oai_list_vector_store_files()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$list_files(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-retrieve_file"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-retrieve_file}{}}}
\subsection{Method \code{retrieve_file()}}{
Retrieve a file in the vector store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$retrieve_file(file_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{Character. The ID of the file to be added to the vector store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-delete_file"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-delete_file}{}}}
\subsection{Method \code{delete_file()}}{
Delete a file in the vector store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$delete_file(file_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{Character. The ID of the file to be added to the vector store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-search"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-search}{}}}
\subsection{Method \code{search()}}{
Search the vector store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$search(query, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{Character. The query string for the search.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}

\item{\code{filters}}{List. Optional. A filter to apply based on file attributes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-create_file_batch"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-create_file_batch}{}}}
\subsection{Method \code{create_file_batch()}}{
Create a file batch. The \code{...} argument is passed to \code{oai_create_vector_store_file_batch()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$create_file_batch(file_ids, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_ids}}{Character. Optional. A vector of file IDs to include in the vector store.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A VectorStoreFilesBatch R6 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-retrieve_file_batch"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-retrieve_file_batch}{}}}
\subsection{Method \code{retrieve_file_batch()}}{
Retrieve a file batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$retrieve_file_batch(batch_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_id}}{Character. The ID of the batch to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-cancel_file_batch"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-cancel_file_batch}{}}}
\subsection{Method \code{cancel_file_batch()}}{
Cancel a file batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$cancel_file_batch(batch_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_id}}{Character. The ID of the batch to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-list_files_in_a_batch"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-list_files_in_a_batch}{}}}
\subsection{Method \code{list_files_in_a_batch()}}{
List files in a batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$list_files_in_a_batch(batch_id = batch_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_id}}{Character. The ID of the batch to retrieve.}

\item{\code{...}}{Additional arguments to be passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-print"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-print}{}}}
\subsection{Method \code{print()}}{
Print the VectorStore object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
