% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_store_files.R
\name{VectorStoreFile}
\alias{VectorStoreFile}
\title{VectorStoreFile R6 class}
\description{
VectorStoreFile R6 class

VectorStoreFile R6 class
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{VectorStoreFile}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Character. The ID of the vector store file.}

\item{\code{attributes}}{Named list. A list of attributes associated with the vector store file.}

\item{\code{created_at}}{POSIXct. The time the vector store file was created.}

\item{\code{status}}{Character. The status of the vector store file. One of "in_progress", "completed", "failed", "cancelled".}

\item{\code{chunking_strategy}}{List. The strategy used to chunk the data.}

\item{\code{usage_bytes}}{Integer. The number of bytes used by the vector store file.}

\item{\code{vector_store_id}}{Character. The ID of the vector store that the file belongs to.}

\item{\code{last_error}}{Character. The last error that occurred while processing the file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStoreFile-new}{\code{VectorStoreFile$new()}}
\item \href{#method-VectorStoreFile-retrieve}{\code{VectorStoreFile$retrieve()}}
\item \href{#method-VectorStoreFile-update}{\code{VectorStoreFile$update()}}
\item \href{#method-VectorStoreFile-delete}{\code{VectorStoreFile$delete()}}
\item \href{#method-VectorStoreFile-print}{\code{VectorStoreFile$print()}}
\item \href{#method-VectorStoreFile-clone}{\code{VectorStoreFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new VectorStoreFile object.
Either provide a vector_store_file_id or a vector_store_id together with a file_id or a path to a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$new(
  vector_store_file_id = NULL,
  vector_store_id = NULL,
  file_id = NULL,
  path = NULL,
  resp = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector_store_file_id}}{Character. The ID of the vector store file.}

\item{\code{vector_store_id}}{Character. The ID of the vector store that the file belongs to.}

\item{\code{file_id}}{Character. The ID of the file that the vector store file is associated with.}

\item{\code{path}}{Character. The path to the file that the vector store file is associated with.}

\item{\code{resp}}{List. The response object from the OpenAI API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the vector store file from the OpenAI API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-update"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-update}{}}}
\subsection{Method \code{update()}}{
Update the vector store file in the OpenAI API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$update(attributes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attributes}}{Named list. A list of attributes to update, with a maximum of 16 key-value pairs. Keys are strings with a maximum length of 64 characters. Values are strings with a maximum length of 512 characters, booleans, or numbers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-delete"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the vector store file from the OpenAI API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-print"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-print}{}}}
\subsection{Method \code{print()}}{
Print the vector store file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
