% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_store_file_batches.R
\name{VectorStoreFilesBatch}
\alias{VectorStoreFilesBatch}
\title{VectorStoreFilesBatch R6 class}
\description{
VectorStoreFilesBatch R6 class

VectorStoreFilesBatch R6 class
}
\section{Super class}{
\code{openaiapi::Utils} -> \code{VectorStoreFilesBatch}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Character. The ID of the file batch.}

\item{\code{attributes}}{List. The attributes of the file batch.}

\item{\code{created_at}}{POSIXct. The date and time the file batch was created.}

\item{\code{vector_store_id}}{Character. The ID of the vector store where the file batch was created.}

\item{\code{status}}{Character. The status of the file batch.}

\item{\code{file_counts}}{List. The number of files in the batch.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStoreFilesBatch-new}{\code{VectorStoreFilesBatch$new()}}
\item \href{#method-VectorStoreFilesBatch-retrieve}{\code{VectorStoreFilesBatch$retrieve()}}
\item \href{#method-VectorStoreFilesBatch-cancel}{\code{VectorStoreFilesBatch$cancel()}}
\item \href{#method-VectorStoreFilesBatch-delete}{\code{VectorStoreFilesBatch$delete()}}
\item \href{#method-VectorStoreFilesBatch-list_files}{\code{VectorStoreFilesBatch$list_files()}}
\item \href{#method-VectorStoreFilesBatch-print}{\code{VectorStoreFilesBatch$print()}}
\item \href{#method-VectorStoreFilesBatch-clone}{\code{VectorStoreFilesBatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a VectorStoreFilesBatch object. The \code{...} argument is passed to the API functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$new(
  vector_store_id = NULL,
  batch_id = NULL,
  file_ids = NULL,
  paths = NULL,
  ...,
  resp = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector_store_id}}{Character. The ID of the vector store where the file batch will be created.}

\item{\code{batch_id}}{Character. The ID of the batch to retrieve.}

\item{\code{file_ids}}{Character. A vector of file IDs to include in the batch.}

\item{\code{paths}}{Character. A vector of file paths to include in the batch.}

\item{\code{...}}{Additional arguments passed to the API functions.}

\item{\code{resp}}{List. The response from the API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve the file batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$retrieve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancel the file batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$cancel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-delete"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the file batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-list_files}{}}}
\subsection{Method \code{list_files()}}{
List all files in the file batch. The \code{...} argument is passed to \code{oai_list_vector_store_files()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$list_files(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-print"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-print}{}}}
\subsection{Method \code{print()}}{
Print the file batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the API functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoreFilesBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoreFilesBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoreFilesBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
