% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{ChatCompletion}
\alias{ChatCompletion}
\title{ChatCompletion R6 class}
\description{
ChatCompletion R6 class

ChatCompletion R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{A unique identifier for the chat completion.}

\item{\code{choices}}{A list of chat completion choices. Can be more than one if n is greater than 1.}

\item{\code{created}}{The Unix timestamp (in seconds) of when the chat completion was created.}

\item{\code{model}}{The model used for the chat completion.}

\item{\code{service_tier}}{The service tier used for processing the request.}

\item{\code{system_fingerprint}}{This fingerprint represents the backend configuration that the model runs with.}

\item{\code{usage}}{Usage statistics for the completion request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChatCompletion-new}{\code{ChatCompletion$new()}}
\item \href{#method-ChatCompletion-get}{\code{ChatCompletion$get()}}
\item \href{#method-ChatCompletion-update}{\code{ChatCompletion$update()}}
\item \href{#method-ChatCompletion-delete}{\code{ChatCompletion$delete()}}
\item \href{#method-ChatCompletion-get_messages}{\code{ChatCompletion$get_messages()}}
\item \href{#method-ChatCompletion-content_text}{\code{ChatCompletion$content_text()}}
\item \href{#method-ChatCompletion-clone}{\code{ChatCompletion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ChatCompletion object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$new(completion_id = NULL, messages = NULL, ..., resp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{completion_id}}{A unique identifier for the chat completion.}

\item{\code{messages}}{A list of messages comprising the conversation so far.}

\item{\code{...}}{Additional parameters passed to API functions}

\item{\code{resp}}{A response object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-get"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-get}{}}}
\subsection{Method \code{get()}}{
Get the chat completion object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-update"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-update}{}}}
\subsection{Method \code{update()}}{
Update the chat completion object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$update(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{List of key-value pairs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-delete"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-delete}{}}}
\subsection{Method \code{delete()}}{
Delete the chat completion object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-get_messages"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-get_messages}{}}}
\subsection{Method \code{get_messages()}}{
Get the messages in the chat completion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$get_messages(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters passed to API functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-content_text"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-content_text}{}}}
\subsection{Method \code{content_text()}}{
Get the messages in the chat completion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$content_text()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
